renamed channel_info to draft_post_header
