+++ app/components/draft/draft.tsx
@@ -7,8 +7,8 @@ import {Keyboard, TouchableHighlight, View} from 'react-native';
 
 import {switchToThread} from '@actions/local/thread';
 import {switchToChannelById} from '@actions/remote/channel';
-import ChannelInfo from '@components/channel_info';
 import DraftPost from '@components/draft/draft_post';
+import ChannelInfo from '@components/draft_post_header';
 import Header from '@components/post_draft/draft_input/header';
 import {Screens} from '@constants';
 import {useServerUrl} from '@context/server';
